refactor: adjust reader directory location
