+++ src/etl/FhirEtlModule.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common'
 
 import { FhirEtlService } from './FhirEtlService'
+import { ReaderModule } from './reader/ReaderModule'
 import { ElasticsearchModule } from '../shared/elasticsearch/ElasticsearchModule'
 import { LoggerModule } from '../shared/logger/LoggerModule'
-import { ReaderModule } from '../shared/reader/ReaderModule'
 
 @Module({
   imports: [LoggerModule, ElasticsearchModule, ReaderModule],
